# lib/api-zod/src/generated/types/friendsResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CLUBSA profile, club, squad, invitation, market, and referral API
 * OpenAPI spec version: 0.1.0
 */
import type { SocialUser } from './socialUser';

export interface FriendsResponse {
  friends: SocialUser[];
  incomingRequests: SocialUser[];
  outgoingRequests: SocialUser[];
}