# lib/api-zod/src/generated/types/friendshipActionResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CLUBSA profile, club, squad, invitation, market, and referral API
 * OpenAPI spec version: 0.1.0
 */
import type { FriendshipActionResponseStatus } from './friendshipActionResponseStatus';

export interface FriendshipActionResponse {
  status: FriendshipActionResponseStatus;
}