# lib/api-zod/src/generated/types/privateMessage.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CLUBSA profile, club, squad, invitation, market, and referral API
 * OpenAPI spec version: 0.1.0
 */

export interface PrivateMessage {
  id: string;
  senderId: string;
  receiverId: string;
  content: string;
  createdAt: Date;
}