# lib/api-zod/src/generated/types/privateMessageInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CLUBSA profile, club, squad, invitation, market, and referral API
 * OpenAPI spec version: 0.1.0
 */

export interface PrivateMessageInput {
  /**
     * @minLength 1
     * @maxLength 255
     */
  receiverId: string;
  /**
     * @minLength 1
     * @maxLength 4000
     */
  content: string;
}