# lib/api-zod/src/generated/types/privateMessageResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CLUBSA profile, club, squad, invitation, market, and referral API
 * OpenAPI spec version: 0.1.0
 */
import type { PrivateMessage } from './privateMessage';

export interface PrivateMessageResponse {
  message: PrivateMessage;
}