# lib/api-zod/src/generated/types/userSearchResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CLUBSA profile, club, squad, invitation, market, and referral API
 * OpenAPI spec version: 0.1.0
 */
import type { SocialUser } from './socialUser';
import type { UserSearchResultRelationshipStatus } from './userSearchResultRelationshipStatus';

export type UserSearchResult = SocialUser & {
  relationshipStatus: UserSearchResultRelationshipStatus;
};